tddgo/hello/hello.go
package main

const chinese = "Chinese"
const french = "French"
const english = "English"
const englishHelloPrefix = "Hello, "
const chineseHelloPrefix = "你好， "
const frenchHelloPrefix = "Bonjour, "

func Hello(name string, language string) string {
	if name == "" {
		name = "World"
	}

	return greetingPrefix(language) + name
}

func greetingPrefix(language string) (prefix string) {
	switch language {
	case chinese:
		prefix = chineseHelloPrefix
	case french:
		prefix = frenchHelloPrefix
	default:
		prefix = englishHelloPrefix
	}
	return
}
